requirements.txt
streamlit==1.35.0
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-text-splitters==0.2.0
langchain-ollama==0.1.0
chromadb==0.4.24
beautifulsoup4==4.12.3
requests==2.31.0
pypdf==4.0.1
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10
langdetect==1.0.9
lxml==5.1.0